split window into a service to break dependency cycle
